ote_cli file naming change
