+++ external/model-preparation-algorithm/tests/ote_cli/test_detection.py
@@ -1,4 +1,4 @@
-"""Tests for MPA Class-Incremental Learning for image classification with OTE CLI"""
+"""Tests for MPA Class-Incremental Learning for object detection with OTE CLI"""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -28,60 +28,59 @@ from ote_cli.utils.tests import (
     nncf_export_testing,
     nncf_eval_testing,
     nncf_eval_openvino_testing,
-    xfail_templates,
 )
 
-# Pre-train w/ 'intel', 'openvino' classes
+# Pre-train w/ 'person' class
 args0 = {
-    '--train-ann-file': '',
-    '--train-data-roots': 'data/text_recognition/initial_data',
-    '--val-ann-file': '',
-    '--val-data-roots': 'data/text_recognition/initial_data',
-    '--test-ann-files': '',
-    '--test-data-roots': 'data/text_recognition/initial_data',
-    '--input': 'data/text_recognition/initial_data/intel',
+    '--train-ann-file': 'data/airport/annotation_person_train.json',
+    '--train-data-roots': 'data/airport/train',
+    '--val-ann-file': 'data/airport/annotation_person_train.json',
+    '--val-data-roots': 'data/airport/train',
+    '--test-ann-files': 'data/airport/annotation_person_train.json',
+    '--test-data-roots': 'data/airport/train',
+    '--input': 'data/airport/train',
     'train_params': [
         'params',
         '--learning_parameters.num_iters',
         '2',
         '--learning_parameters.batch_size',
-        '2',
+        '2'
     ]
 }
 
-# Pre-train w/ 'intel', 'openvino', 'opencv' classes
+# Class-Incremental learning w/ 'vehicle', 'person', 'non-vehicle' classes
 args = {
-    '--train-ann-file': '',
-    '--train-data-roots': 'data/text_recognition/IL_data',
-    '--val-ann-file': '',
-    '--val-data-roots': 'data/text_recognition/IL_data',
-    '--test-ann-files': '',
-    '--test-data-roots': 'data/text_recognition/IL_data',
-    '--input': 'data/text_recognition/IL_data/intel',
+    '--train-ann-file': 'data/airport/annotation_example_train.json',
+    '--train-data-roots': 'data/airport/train',
+    '--val-ann-file': 'data/airport/annotation_example_train.json',
+    '--val-data-roots': 'data/airport/train',
+    '--test-ann-files': 'data/airport/annotation_example_train.json',
+    '--test-data-roots': 'data/airport/train',
+    '--input': 'data/airport/train',
     'train_params': [
         'params',
         '--learning_parameters.num_iters',
         '2',
         '--learning_parameters.batch_size',
-        '4',
+        '2'
     ]
 }
 
 root = '/tmp/ote_cli/'
 ote_dir = os.getcwd()
 
-templates = Registry('external/model-preparation-algorithm').filter(task_type='CLASSIFICATION').templates
+templates = Registry('external/model-preparation-algorithm').filter(task_type='DETECTION').templates
 templates_ids = [template.model_template_id for template in templates]
 
 
-class TestToolsClsClsIncr:
+class TestToolsMPADetection:
     @e2e_pytest_component
     def test_create_venv(self):
         work_dir, _, algo_backend_dir = get_some_vars(templates[0], root)
         create_venv(algo_backend_dir, work_dir)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_train(self, template):
         ote_train_testing(template, root, ote_dir, args0)
         _, template_work_dir, _ = get_some_vars(template, root)
@@ -90,52 +89,52 @@ class TestToolsClsClsIncr:
         ote_train_testing(template, root, ote_dir, args1)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_export(self, template):
         ote_export_testing(template, root)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_eval(self, template):
         ote_eval_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_eval_openvino(self, template):
-        ote_eval_openvino_testing(template, root, ote_dir, args, threshold=0.0)
+        ote_eval_openvino_testing(template, root, ote_dir, args, threshold=0.2)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_demo(self, template):
         ote_demo_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_demo_openvino(self, template):
         ote_demo_openvino_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_deploy_openvino(self, template):
         ote_deploy_openvino_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_eval_deployment(self, template):
         ote_eval_deployment_testing(template, root, ote_dir, args, threshold=0.0)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_demo_deployment(self, template):
         ote_demo_deployment_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_ote_hpo(self, template):
         ote_hpo_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_nncf_optimize(self, template):
         if template.entrypoints.nncf is None:
             pytest.skip("nncf entrypoint is none")
@@ -143,7 +142,7 @@ class TestToolsClsClsIncr:
         nncf_optimize_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_nncf_export(self, template):
         if template.entrypoints.nncf is None:
             pytest.skip("nncf entrypoint is none")
@@ -151,7 +150,7 @@ class TestToolsClsClsIncr:
         nncf_export_testing(template, root)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_nncf_eval(self, template):
         if template.entrypoints.nncf is None:
             pytest.skip("nncf entrypoint is none")
@@ -159,7 +158,7 @@ class TestToolsClsClsIncr:
         nncf_eval_testing(template, root, ote_dir, args, threshold=0.001)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_nncf_eval_openvino(self, template):
         if template.entrypoints.nncf is None:
             pytest.skip("nncf entrypoint is none")
@@ -167,11 +166,11 @@ class TestToolsClsClsIncr:
         nncf_eval_openvino_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_pot_optimize(self, template):
         pot_optimize_testing(template, root, ote_dir, args)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize('template', templates, ids=templates_ids)
     def test_pot_eval(self, template):
         pot_eval_testing(template, root, ote_dir, args)
