+++ external/model-preparation-algorithm/tests/ote_cli/test_iseg.py
@@ -73,7 +73,7 @@ templates = Registry(
 templates_ids = [template.model_template_id for template in templates]
 
 
-class TestToolsInstanceSegmentationClsIncr:
+class TestToolsMPAInstanceSegmentation:
     @e2e_pytest_component
     def test_create_venv(self):
         work_dir, _, algo_backend_dir = get_some_vars(templates[0], root)
